+++ src/utils.jl
@@ -1,20 +1,23 @@
-using Preferences
+import Preferences
 
-const DEBUG = @load_preference("use_debug", false)
+const DEBUG = Preferences.@load_preference("use_debug", false)
 
 """
-debug_mode(enable = true)
+    Ferrite.debug_mode(; enable=true)
 
-Helper to turn on or off debug helpers in Ferrite.
+Helper to turn on (`enable=true`) or off (`enable=false`) debug expressions in Ferrite.
+
+Debug mode influences `Ferrite.@debug expr`: when debug mode is enabled, `expr` is
+evaluated, and when debug mode is disabled `expr` is ignored.
 """
-function debug_mode(enable = true)
-    @set_preferences!("use_debug" => enable)
-    DEBUG_STR = DEBUG ? "ON" : "OFF"
-    NEW_DEBUG_STR = enable ? "ON" : "OFF"
-    if DEBUG != enable
-        @info("Toggling debug mode from $DEBUG_STR to $NEW_DEBUG_STR. Restart your Julia session for this change to take effect!")
+function debug_mode(; enable = true)
+    if DEBUG == enable == true
+        @info "Debug mode already enabled."
+    elseif DEBUG == enable == false
+        @info "Debug mode already disabled."
     else
-        @info("Debug mode already $DEBUG_STR.")
+        Preferences.@set_preferences!("use_debug" => enable)
+        @info "Debug mode $(enable ? "en" : "dis")abled. Restart the Julia session for this change to take effect!"
     end
 end